# Make `load_geometry` type stable