`load_geometry(file)`, where `file` is a ~7.4MB large (~150k faces) stl-file:
 ### `main`
```julia
julia> @benchmark load_geometry($file)
BenchmarkTools.Trial: 4 samples with 1 evaluation.
 Range (min … max):  1.529 s …   1.596 s  ┊ GC (min … max): 1.07% … 3.03%
 Time  (median):     1.571 s              ┊ GC (median):    1.03%
 Time  (mean ± σ):   1.567 s ± 30.554 ms  ┊ GC (mean ± σ):  1.51% ± 1.01%

  █                      █                        █       █  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁█ ▁
  1.53 s         Histogram: frequency by time         1.6 s <

 Memory estimate: 529.69 MiB, allocs estimate: 11601626.
```

### This PR
```julia
julia> @benchmark load_geometry($file)
BenchmarkTools.Trial: 72 samples with 1 evaluation.
 Range (min … max):  61.315 ms … 113.754 ms  ┊ GC (min … max): 0.00% … 7.92%
 Time  (median):     69.208 ms               ┊ GC (median):    4.10%
 Time  (mean ± σ):   69.877 ms ±   6.842 ms  ┊ GC (mean ± σ):  3.92% ± 1.99%

                ▅█ ▂▂        ▂         ▂      ▂  ▅              
  ▅▅▅▁█▁▁▁▅█▅▅████▅██▅▁▅▅▁█▁▁█▁▁▅▅█▅█▁▅███▅▅█▁█▅▁█▁▁▅█▁▅▅▁▁▁▁▅ ▁
  61.3 ms         Histogram: frequency by time         78.6 ms <

 Memory estimate: 89.37 MiB, allocs estimate: 191.
```